Move ChromeAdapter class to it's own file
