+++ src/background/dev-tools-chrome-adapter.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BrowserAdapter, ChromeAdapter } from './browser-adapters/browser-adapter';
+import { BrowserAdapter } from './browser-adapters/browser-adapter';
+import { ChromeAdapter } from './browser-adapters/chrome-adapter';
 
 export interface DevToolsChromeAdapter extends BrowserAdapter {
     getInspectedWindowTabId(): number;
