+++ src/tests/unit/tests/common/notification-creator.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
-import { ChromeAdapter } from '../../../../background/browser-adapters/browser-adapter';
+import { BrowserAdapter } from '../../../../background/browser-adapters/browser-adapter';
 import {
     VisualizationConfiguration,
     VisualizationConfigurationFactory,
@@ -11,7 +11,7 @@ import { NotificationCreator } from '../../../../common/notification-creator';
 import { VisualizationType } from '../../../../common/types/visualization-type';
 
 describe('NotificationCreator', () => {
-    let browserAdapterMock: IMock<ChromeAdapter>;
+    let browserAdapterMock: IMock<BrowserAdapter>;
     let configFactoryMock: IMock<VisualizationConfigurationFactory>;
     let getNotificationMessageMock: IMock<(selectorMap, key) => string>;
     let testObject: NotificationCreator;
@@ -19,7 +19,7 @@ describe('NotificationCreator', () => {
     const visualizationType: VisualizationType = -1;
 
     beforeEach(() => {
-        browserAdapterMock = Mock.ofType(ChromeAdapter, MockBehavior.Strict);
+        browserAdapterMock = Mock.ofType<BrowserAdapter>(null, MockBehavior.Strict);
         configFactoryMock = Mock.ofType(VisualizationConfigurationFactory, MockBehavior.Strict);
         getNotificationMessageMock = Mock.ofInstance(selector => null);
         testObject = new NotificationCreator(browserAdapterMock.object, configFactoryMock.object);
