+++ src/tests/unit/tests/popup/components/popup-view.test.tsx
@@ -3,7 +3,8 @@
 import { shallow } from 'enzyme';
 import * as React from 'react';
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
-import { ChromeAdapter } from '../../../../../background/browser-adapters/browser-adapter';
+
+import { BrowserAdapter } from '../../../../../background/browser-adapters/browser-adapter';
 import { NewTabLink } from '../../../../../common/components/new-tab-link';
 import { DropdownClickHandler } from '../../../../../common/dropdown-click-handler';
 import { StoreActionMessageCreatorImpl } from '../../../../../common/message-creators/store-action-message-creator-impl';
@@ -27,7 +28,7 @@ import { BaseDataBuilder } from '../../../common/base-data-builder';
 import { IsSameObject } from '../../../common/typemoq-helper';
 
 describe('PopupView', () => {
-    const browserAdapterMock = Mock.ofType(ChromeAdapter);
+    const browserAdapterMock = Mock.ofType<BrowserAdapter>();
     const launchPanelStateStoreState: LaunchPanelStoreData = {
         launchPanelType: LaunchPanelType.LaunchPad,
     };
